backend/requirements.txt
cemotion>=2.0.0
flask>=3.0.0
psutil>=5.9.0
addict>=2.4.0
modelscope>=1.28.2
datasets==3.6.0
Pillow>=10.0.0
simplejson>=3.19.0
sortedcontainers>=2.4.0

# 鉴权系统依赖
Flask-SQLAlchemy>=3.0.0
Flask-JWT-Extended>=4.6.0
Flask-RESTX>=1.3.0
Flask-Mail>=0.9.1
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
PyJWT>=2.8.0
bcrypt>=4.1.0
marshmallow>=3.20.0
sqlalchemy-utils>=0.41.0
psycopg2-binary>=2.9.7
gunicorn>=21.2.0
